requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.59.0
PyGithub>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
groq>=0.4.0
google-generativeai>=0.4.0
streamlit>=1.30.0
pandas<3.0.0
